add vertical preview class
