+++ embulk-core/src/main/java/org/embulk/command/TablePrinter.java
@@ -6,59 +6,20 @@ import java.util.Locale;
 import java.io.PrintStream;
 import java.io.Flushable;
 import java.io.IOException;
+import org.embulk.config.ModelManager;
 
-class TablePrinter
+class TablePrinter extends PreviewPrinter
 {
-    private static final int SAMPLES = 10;
 
-    private PrintStream out;
-
-    private String[] header;
-
-    private List<Object[]> samples;
     private String format;
     private String border;
 
-    public TablePrinter(PrintStream out, String... header)
+    public TablePrinter(PrintStream out, ModelManager model, String... header)
     {
-        this.out = out;
-        this.header = header;
-        this.samples = new ArrayList<Object[]>(SAMPLES);
-        samples.add(header);
+        super(out, model, header);
     }
 
-    public void add(Object... values) throws IOException
-    {
-        int min = header.length < values.length ? header.length : values.length;
-        Object[] cols = new Object[header.length];
-        for(int i=0; i < min; i++) {
-            if(values[i] == null) {
-                cols[i] = "";
-            } else {
-                cols[i] = valueToString(values[i]);
-            }
-        }
-        for(int i=min; i < header.length; i++) {
-            cols[i] = "";
-        }
-
-        if(samples == null) {
-            out.format(format, cols);
-            return;
-        }
-
-        samples.add(cols);
-        if(samples.size() < SAMPLES) {
-            return;
-        }
-    }
-
-    protected String valueToString(Object obj)
-    {
-        return obj.toString();
-    }
-
-    private void flushSamples()
+    protected void flushSamples()
     {
         StringBuilder borderBuilder = new StringBuilder();
 
@@ -92,28 +53,10 @@ class TablePrinter
             }
         }
 
+        out.println(border);
+
         this.samples = null;
     }
 
-    private int maxLengthInColumn(int i)
-    {
-        int max = 0;
-        for(Object[] cols : samples) {
-            String s = (String) cols[i];
-            int len = (s == null) ? 0 : s.length();
-            if(max < len) {
-                max = len;
-            }
-        }
-        return max;
-    }
-
-    public void finish() throws IOException
-    {
-        if(!samples.isEmpty()) {
-            flushSamples();
-        }
-        out.println(border);
-    }
 }
